+++ src/main/java/nextstep/ladder/model/Ladder.java
@@ -4,22 +4,21 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Ladder {
-    private final List<Lines> lines;
+	private final List<Lines> lines;
 
-    public Ladder(final int numberOfFloors, final int numberOfPlayers) {
-        List<Lines> result = new ArrayList<>();
-        for (int i = 0; i < numberOfFloors; i++) {
-            result.add(new Lines(numberOfPlayers));
-        }
-        this.lines = result;
-    }
+	public Ladder(final int numberOfFloors, final int numberOfPlayers) {
+		List<Lines> result = new ArrayList<>();
+		for (int i = 0; i < numberOfFloors; i++) {
+			result.add(new Lines(numberOfPlayers));
+		}
+		this.lines = result;
+	}
 
+	public int numberOfFloors() {
+		return lines.size();
+	}
 
-    public int numberOfFloors() {
-        return lines.size();
-    }
-
-    public List<Lines> getLines() {
-        return lines;
-    }
+	public List<Lines> getLines() {
+		return lines;
+	}
 }
