+++ src/main/java/nextstep/ladder/view/Output.java
@@ -8,16 +8,16 @@ import nextstep.ladder.model.Players;
 import java.util.List;
 
 public class Output {
-    private Output() {
-    }
+	private Output() {
+	}
 
-    public static void outputResult(Ladder ladder, Players players){
-        System.out.println(players);
-        List<Lines> lines = ladder.getLines();
-        for (Lines line : lines) {
-            System.out.print("     |");
-            System.out.print(line);
-            System.out.println();
-        }
-    }
+	public static void outputResult(Ladder ladder, Players players) {
+		System.out.println(players);
+		List<Lines> lines = ladder.getLines();
+		for (Lines line : lines) {
+			System.out.print("     |");
+			System.out.print(line);
+			System.out.println();
+		}
+	}
 }
