+++ src/main/java/nextstep/optional/ComputerStore.java
@@ -4,23 +4,23 @@ import nextstep.optional.Computer.Soundcard;
 import nextstep.optional.Computer.USB;
 
 public class ComputerStore {
-    public static final String UNKNOWN_VERSION = "UNKNOWN";
+	public static final String UNKNOWN_VERSION = "UNKNOWN";
 
-    public static String getVersion(Computer computer) {
-        String version = UNKNOWN_VERSION;
-        if (computer != null) {
-            Soundcard soundcard = computer.getSoundcard();
-            if (soundcard != null) {
-                USB usb = soundcard.getUsb();
-                if (usb != null) {
-                    version = usb.getVersion();
-                }
-            }
-        }
-        return version;
-    }
+	public static String getVersion(Computer computer) {
+		String version = UNKNOWN_VERSION;
+		if (computer != null) {
+			Soundcard soundcard = computer.getSoundcard();
+			if (soundcard != null) {
+				USB usb = soundcard.getUsb();
+				if (usb != null) {
+					version = usb.getVersion();
+				}
+			}
+		}
+		return version;
+	}
 
-    public static String getVersionOptional(Computer computer) {
-        return null;
-    }
+	public static String getVersionOptional(Computer computer) {
+		return null;
+	}
 }
