+++ src/main/java/nextstep/optional/Expression.java
@@ -3,22 +3,22 @@ package nextstep.optional;
 import java.util.Arrays;
 
 enum Expression {
-    PLUS("+"), MINUS("-"), TIMES("*"), DIVIDE("/");
+	PLUS("+"), MINUS("-"), TIMES("*"), DIVIDE("/");
 
-    private String expression;
+	private String expression;
 
-    Expression(String expression) {
-        this.expression = expression;
-    }
+	Expression(String expression) {
+		this.expression = expression;
+	}
 
-    private static boolean matchExpression(Expression e, String expression) {
-        return expression.equals(e.expression);
-    }
+	private static boolean matchExpression(Expression e, String expression) {
+		return expression.equals(e.expression);
+	}
 
-    static Expression of(String expression) {
-        return Arrays.stream(values())
-            .filter(value -> matchExpression(value, expression))
-            .findAny()
-            .orElseThrow(() -> new IllegalArgumentException(String.format("%s는 사칙연산에 해당하지 않는 표현식입니다.", expression)));
-    }
+	static Expression of(String expression) {
+		return Arrays.stream(values())
+			.filter(value -> matchExpression(value, expression))
+			.findAny()
+			.orElseThrow(() -> new IllegalArgumentException(String.format("%s는 사칙연산에 해당하지 않는 표현식입니다.", expression)));
+	}
 }
