+++ src/main/java/nextstep/optional/Users.java
@@ -4,18 +4,19 @@ import java.util.Arrays;
 import java.util.List;
 
 public class Users {
-    static final User DEFAULT_USER = new User("codesquad", 100);
+	static final User DEFAULT_USER = new User("codesquad", 100);
 
-    List<User> users = Arrays.asList(
-            new User("crong", 35),
-            new User("pobi", 30),
-            new User("jk", 40),
-            new User("honux", 45));
+	List<User> users = Arrays.asList(
+		new User("crong", 35),
+		new User("pobi", 30),
+		new User("jk", 40),
+		new User("honux", 45)
+	);
 
-    User getUser(String name) {
-        return users.stream()
-            .filter(user -> user.matchName(name))
-            .findAny()
-            .orElse(DEFAULT_USER);
-    }
+	User getUser(String name) {
+		return users.stream()
+			.filter(user -> user.matchName(name))
+			.findAny()
+			.orElse(DEFAULT_USER);
+	}
 }
