+++ src/test/java/nextstep/ladder/model/PlayerTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.model;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -10,8 +10,7 @@ class PlayerTest {
 
 	@Test
 	void 사용자의_이름은_5자를_넘길수_없다() {
-		assertThatIllegalArgumentException()
-			.isThrownBy(() -> new Player("abcdef"))
+		assertThatIllegalArgumentException().isThrownBy(() -> new Player("abcdef"))
 			.withMessage("사용자의 이름은 5자를 넘을 수 없습니다");
 	}
